requirements.txt
torch==2.1.0
transformers==4.41.1
tokenizers==0.19.1
sentencepiece==0.1.99
pyarrow==14.0.2
accelerate==0.29.0
pydantic<2,>=1
markdown2[all]
numpy==1.24.3
scikit-learn==1.2.2
requests
httpx==0.24.0
uvicorn
fastapi
einops==0.6.1
einops-exts==0.0.4
timm==0.6.13
tiktoken
transformers_stream_generator==0.0.4
scipy
deepspeed==0.14.0
pandas
torchaudio
xformers
pillow==10.3.0